requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
abseil-cpp=20211102.0=h27087fc_1
absl-py=1.3.0=py39h06a4308_0
aiohttp=3.8.3=py39h5eee18b_0
aiosignal=1.3.1=pyhd8ed1ab_0
anyio=3.5.0=py39h06a4308_0
argon2-cffi=21.3.0=pyhd3eb1b0_0
argon2-cffi-bindings=21.2.0=py39h7f8727e_0
arrow-cpp=8.0.0=py39h60b952e_0
asttokens=2.0.5=pyhd3eb1b0_0
async-timeout=4.0.2=pyhd8ed1ab_0
attrs=22.1.0=pyh71513ae_1
aws-c-common=0.4.57=he6710b0_1
aws-c-event-stream=0.1.6=h2531618_5
aws-checksums=0.1.9=he6710b0_0
aws-sdk-cpp=1.8.185=hce553d0_0
backcall=0.2.0=pyhd3eb1b0_0
beautifulsoup4=4.11.1=py39h06a4308_0
blas=1.0=mkl
bleach=4.1.0=pyhd3eb1b0_0
boost-cpp=1.78.0=he72f1d9_0
bottleneck=1.3.5=py39h7deecbd_0
brotli=1.0.9=h166bdaf_7
brotli-bin=1.0.9=h166bdaf_7
brotlipy=0.7.0=py39h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
c-ares=1.18.1=h7f98852_0
ca-certificates=2022.10.11=h06a4308_0
certifi=2022.9.24=py39h06a4308_0
cffi=1.15.1=py39h5eee18b_2
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.3=py39hf3d152e_1
colorama=0.4.6=pyhd8ed1ab_0
cryptography=38.0.1=py39h9ce1e76_0
cuda=11.7.1=0
cuda-cccl=11.7.91=0
cuda-command-line-tools=11.7.1=0
cuda-compiler=11.7.1=0
cuda-cudart=11.7.99=0
cuda-cudart-dev=11.7.99=0
cuda-cuobjdump=11.7.91=0
cuda-cupti=11.7.101=0
cuda-cuxxfilt=11.7.91=0
cuda-demo-suite=11.8.86=0
cuda-documentation=11.8.86=0
cuda-driver-dev=11.7.99=0
cuda-gdb=11.8.86=0
cuda-libraries=11.7.1=0
cuda-libraries-dev=11.7.1=0
cuda-memcheck=11.8.86=0
cuda-nsight=11.8.86=0
cuda-nsight-compute=11.8.0=0
cuda-nvcc=11.7.99=0
cuda-nvdisasm=11.8.86=0
cuda-nvml-dev=11.7.91=0
cuda-nvprof=11.8.87=0
cuda-nvprune=11.7.91=0
cuda-nvrtc=11.7.99=0
cuda-nvrtc-dev=11.7.99=0
cuda-nvtx=11.7.91=0
cuda-nvvp=11.8.87=0
cuda-runtime=11.7.1=0
cuda-sanitizer-api=11.8.86=0
cuda-toolkit=11.7.1=0
cuda-tools=11.7.1=0
cuda-visual-tools=11.7.1=0
cudatoolkit=11.7.0=hd8887f6_10
dataclasses=0.8=pyhc8e2a94_3
datasets=2.7.1=pyhd8ed1ab_0
debugpy=1.5.1=py39h295c915_0
decorator=5.1.1=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
dill=0.3.6=pyhd8ed1ab_1
emoji=2.2.0=pyhd8ed1ab_0
entrypoints=0.4=py39h06a4308_0
evaluate=0.3.0=pypi_0
executing=0.8.3=pyhd3eb1b0_0
ffmpeg=4.3=hf484d3e_0
fftw=3.3.10=nompi_h77c792f_102
filelock=3.8.2=pyhd8ed1ab_0
flit-core=3.6.0=pyhd3eb1b0_0
freetype=2.12.1=h4a9f257_0
frozenlist=1.3.3=py39h5eee18b_0
fsspec=2022.11.0=pyhd8ed1ab_0
gds-tools=1.4.0.31=0
gflags=2.2.2=he1b5a44_1004
giflib=5.2.1=h7b6447c_0
glog=0.6.0=h6f12383_0
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
grpc-cpp=1.46.1=h33aed49_0
huggingface_hub=0.11.1=pyhd8ed1ab_0
icu=70.1=h27087fc_0
idna=3.4=py39h06a4308_0
importlib-metadata=5.1.0=pyha770c72_0
importlib_metadata=5.1.0=hd8ed1ab_0
intel-openmp=2021.4.0=h06a4308_3561
ipykernel=6.15.2=py39h06a4308_0
ipython=8.6.0=py39h06a4308_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=7.6.5=pyhd3eb1b0_1
jedi=0.18.1=py39h06a4308_1
jinja2=3.1.2=py39h06a4308_0
joblib=1.2.0=pyhd8ed1ab_0
jpeg=9e=h7f8727e_0
jsonschema=4.16.0=py39h06a4308_0
jupyter_client=7.4.7=py39h06a4308_0
jupyter_core=4.11.2=py39h06a4308_0
jupyter_server=1.18.1=py39h06a4308_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_widgets=1.0.0=pyhd3eb1b0_1
keyutils=1.6.1=h166bdaf_0
krb5=1.19.3=h3790be6_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libbrotlicommon=1.0.9=h166bdaf_7
libbrotlidec=1.0.9=h166bdaf_7
libbrotlienc=1.0.9=h166bdaf_7
libcublas=11.11.3.6=0
libcublas-dev=11.11.3.6=0
libcufft=10.9.0.58=0
libcufft-dev=10.9.0.58=0
libcufile=1.4.0.31=0
libcufile-dev=1.4.0.31=0
libcurand=10.3.0.86=0
libcurand-dev=10.3.0.86=0
libcurl=7.86.0=h91b91d3_0
libcusolver=11.4.1.48=0
libcusolver-dev=11.4.1.48=0
libcusparse=11.7.5.86=0
libcusparse-dev=11.7.5.86=0
libdeflate=1.8=h7f8727e_5
libedit=3.1.20210910=h7f8727e_0
libev=4.33=h516909a_1
libevent=2.1.12=h8f2d780_0
libffi=3.4.2=h6a678d5_6
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=12.2.0=h69a702a_19
libgfortran5=12.2.0=h337968e_19
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.2=h7f8727e_0
libnghttp2=1.46.0=hce63b2e_0
libnpp=11.8.0.86=0
libnpp-dev=11.8.0.86=0
libnvjpeg=11.9.0.86=0
libnvjpeg-dev=11.9.0.86=0
libpng=1.6.37=hbc83047_0
libprotobuf=3.20.1=h4ff587b_0
libsodium=1.0.18=h7b6447c_0
libssh2=1.10.0=ha56f1ee_2
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.16.0=h27cfd23_0
libthrift=0.15.0=hcc01f38_0
libtiff=4.4.0=hecacb30_2
libunistring=0.9.10=h27cfd23_0
libwebp=1.2.4=h11a3e52_0
libwebp-base=1.2.4=h5eee18b_0
lz4-c=1.9.3=h295c915_1
markupsafe=2.1.1=py39h7f8727e_0
matplotlib-inline=0.1.6=py39h06a4308_0
mistune=0.8.4=py39h27cfd23_1000
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py39h7f8727e_0
mkl_fft=1.3.1=py39hd3c417c_0
mkl_random=1.2.2=py39h51133e4_0
multidict=6.0.2=py39hb9d737c_1
multiprocess=0.70.14=py39h06a4308_0
nbclassic=0.4.8=py39h06a4308_0
nbclient=0.5.13=py39h06a4308_0
nbconvert=6.4.4=py39h06a4308_0
nbformat=5.5.0=py39h06a4308_0
ncurses=6.3=h5eee18b_3
nest-asyncio=1.5.5=py39h06a4308_0
nettle=3.7.3=hbbd107a_1
nltk=3.7=pyhd3eb1b0_0
notebook=6.5.2=py39h06a4308_0
notebook-shim=0.2.2=py39h06a4308_0
nsight-compute=2022.3.0.22=0
numexpr=2.8.4=py39he184ba9_0
numpy=1.23.4=py39h14f4228_0
numpy-base=1.23.4=py39h31eccc5_0
openh264=2.1.1=h4ff587b_0
openssl=1.1.1s=h7f8727e_0
orc=1.7.4=h07ed6aa_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.5.1=py39h417a72b_0
pandocfilters=1.5.0=pyhd3eb1b0_0
parso=0.8.3=pyhd3eb1b0_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=9.2.0=py39hace64e9_1
pip=22.2.2=py39h06a4308_0
portalocker=2.6.0=py39hf3d152e_1
prometheus_client=0.14.1=py39h06a4308_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0
protobuf=3.20.1=py39h295c915_0
psutil=5.9.0=py39h5eee18b_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pure_eval=0.2.2=pyhd3eb1b0_0
pyarrow=8.0.0=py39h992f0b0_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.11.2=pyhd3eb1b0_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.9=py39h06a4308_0
pyrsistent=0.18.0=py39heee7806_0
pysocks=1.7.1=py39h06a4308_0
python=3.9.15=h7a1cb2a_2
python-dateutil=2.8.2=pyhd3eb1b0_0
python-fastjsonschema=2.16.2=py39h06a4308_0
python-xxhash=3.0.0=py39hb9d737c_1
python_abi=3.9=2_cp39
pytorch=1.13.0=py3.9_cuda11.7_cudnn8.5.0_0
pytorch-cuda=11.7=h67b0de4_0
pytorch-mutex=1.0=cuda
pytz=2022.1=py39h06a4308_0
pyyaml=6.0=py39hb9d737c_4
pyzmq=23.2.0=py39h6a678d5_0
re2=2022.04.01=h27087fc_0
readline=8.2=h5eee18b_0
regex=2022.7.9=py39h5eee18b_0
requests=2.28.1=py39h06a4308_0
responses=0.18.0=pyhd8ed1ab_0
rouge=1.0.0=py38_0
rouge-score=0.1.2=pypi_0
sacrebleu=2.0.0=pyhd3eb1b0_1
sacremoses=0.0.53=pyhd8ed1ab_0
scikit-learn=1.1.3=py39h6a678d5_0
scipy=1.9.3=py39h14f4228_0
send2trash=1.8.0=pyhd3eb1b0_1
sentencepiece=0.1.96=py39hf939315_1
setuptools=65.5.0=py39h06a4308_0
six=1.16.0=pyhd3eb1b0_1
snappy=1.1.9=hbd366e4_1
sniffio=1.2.0=py39h06a4308_1
soupsieve=2.3.2.post1=py39h06a4308_0
sqlite=3.40.0=h5082296_0
stack_data=0.2.0=pyhd3eb1b0_0
tabulate=0.9.0=pyhd8ed1ab_1
terminado=0.13.1=py39h06a4308_0
testpath=0.6.0=py39h06a4308_0
threadpoolctl=3.1.0=pyh8a188c0_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.11.4=py39h3dcd8bd_1
torchaudio=0.13.0=py39_cu117
torchvision=0.14.0=py39_cu117
tornado=6.2=py39h5eee18b_0
tqdm=4.64.1=pyhd8ed1ab_0
traitlets=5.1.1=pyhd3eb1b0_0
transformers=4.24.0=pyhd8ed1ab_0
typing=3.10.0.0=pyhd8ed1ab_0
typing-extensions=4.4.0=py39h06a4308_0
typing_extensions=4.4.0=py39h06a4308_0
tzdata=2022g=h04d1e81_0
urllib3=1.26.12=py39h06a4308_0
utf8proc=2.6.1=h27cfd23_0
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py39h06a4308_1
websocket-client=0.58.0=py39h06a4308_4
wheel=0.37.1=pyhd3eb1b0_0
widgetsnbextension=3.5.2=py39h06a4308_0
xxhash=0.8.0=h7f98852_3
xz=5.2.8=h5eee18b_0
yaml=0.2.5=h7f98852_2
yarl=1.8.1=py39h5eee18b_0
zeromq=4.3.4=h2531618_0
zipp=3.11.0=pyhd8ed1ab_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.2=ha4553b6_0
